First draft of new permission system.
This update is a first step conversion towards a highly parameterized permission
system for fields. A conversion function for both fields and documents is currently
put together matching the previous logics form.

The previous form had a bug in how it converts elements for JSON. This version
still has that bug. I will be fixing that next.

It is likely I will change the name of `safer_fun` to something more specific to
documents, since that's when this function is applied.