Built the permission system around a single function, make_safe, that takes parameters for common permission needs. A field list can be provided for either a white or black list and transformation functions may also be provided. The options are currently Python structures or JSON safe forms via `to_python()` and `to_json()`.
make_safe is a recursive function that loops around the input parameters until all
structures are covered. It either calls `for_*()` on a field or it unpacks a Document
into either json or python by calling `to_*()` on the document.

Documents are unpacked with `to_*()` functions while fields are readied for some
serialization via `for_*()` functions where '*' represents Python or JSON.