Completed work overhauling the metaclass structures. I first mocked out the gist of the design here: https://gist.github.com/2146589
The high-level is that a `Document` instance can define a class `Meta` to set options
for the class's behavior, such as which field to use for an ID and whether or not the
class is a Mixin.

A user now writes code like:

    class SimpleDoc(Document):
        title = StringField(max_length=40)
        num = IntField()
        class Meta:
            id_field = ObjectIdField

The `Meta` class is then parsed into `_options`, which we see attached to `SimpleDoc`
as `SimpleDoc._options`. The `_options` structure is an instance, or subclass, of
`DocOptions`. (I should probably change that to `DocumentOptions`).

The documentation will be updated soon and will explain all the configuration options
available to document creators.