This commit marks a rename of both DictShield and many of it's components. The main idea is to create better focus in the name of the project and the way users think about it's uses.
Documents are current renamed to Models.
Fields, as in the classes representing a type, are now called Types.

A field is a type instance contained in a model, designated by a name.

This commit represents the first wave of many name changes to come that will solidify
the concepts at work in a user's mind.