Heavy reworking of the metaclass system.
The concept of a Mixin has been removed.  It is replaced instead with simple inheritance.  The concept of a Mixin in previous releases pertained to how the serialization of the class structure was represented.  That component of the system has been removed as it has not proven to be useful.

The class used to configure the meta system is has been renamed from 'Meta' to 'Options'.  This name is what users of the framework use to configure their Models, so calling it 'Options' seems more informative.

The fields `_public_fields` and `_private_fields` are now on the Options class, which now also sets them to be instance level variables instead of class level variables.  I think this will make some folks out there quite happy.  :)