The whitelist and blacklist behavior is finally abstracted and is now a clean system.  The idea is there are now functions called `whitelist` and `blacklist`.
You create one like this:

    blacklist(['field_name', 'other_field'])

The `blacklist` and `whitelist` functions create a new function which
is passed into `apply_shape` as the `gottago` argument.  This function
simply looks at each field during serialization and decides whether or
not the field should be evicted.

That function is what is returned by `blacklist` and `whitelist`.  No
object is used because functions are better.  :)