Migrated the concepts of whitelist and blacklist under the Options for model definitions.
This makes the configuration stored on model instances.

The other change this commit brings is that permissions are now completely
configurable.  The name of groups is a key in a map where the value is a
function that operates as a blacklist or whitelist.  In fact, the functions
`whitelist` and `blacklist` are written such that they themselves create the
correct function to use in `apply_shape` for the `gottago` argument.

In short, you define a class that looks like this, which I find MUCH
cleaner and MUCH more configurable.

    class User(Model):
        secret = MD5Type()
        name = StringType(required=True, max_length=50)
        bio = StringType(max_length=100)

        class Options:
            roles = {
                'owner': blacklist([]),
                'public': whitelist(['name', 'bio']),
            }