This commit puts in place the last pieces of the new the validation system.
There are now three types of validation oriented responses, TypeResult,
FieldResult, and ModelResult.

A TypeResult respresents the result of a calculation at the Type level.  If
a value cannot be coerced into the type, for example.

A FieldResult represents something that happens at the level of an instance of
a type, aka a field.  If one is configured badly, etc.

A ModelResult carries information about the integrity of a model's design.

All of the validation functions at the field level and model level now return
one or more of those values.

Validation on a model is expanded into validation across all of the fields, or
across an expanded `ModelType` and the return values from the fields, which are
FieldResults, are processed into ModelType results to be returned from models.

This mostly completely my idea for a fully exposed validation tree of structures
which can be used to produce error reports as easily as html forms.

After all this work I am really pleased with the namedtuple approach.