Removed the binding of fields to a model
Originally we wanted to keep data on the field object. This forced us
to bind (i.e. create a new instance of the field for every created instance).

Today, we realized that we aren't keeping data on the field object so
bind is just an unecessary object creation.

Fields must therefore be stateless (i.e. each instance of a model must not
change the state of a field) as they are shared across every instance of
a model.