There are two things going in this changeset together.  My mind wandered a a little.
First, a function called `allow_none` has been defined to consolidate where
this logic lives in a single function.  The function takes two arguments, a
class and a field.  It sets it's default value to the value found on the class
but checks the field to see if an override was specified.

Second, the `include_serializables` flag has been removed.  This simplifies
the interface to `to_primitive` and centralizes the method used for serializing
a field, or not serializing a field, in the `Role` system which makes it easier
to reason about.