The commit captures the latest version of the docs.
While writing them I realized there is more work to be done for Schematics. For example, it is not clear how to extract the Python native types after they have been coerced.

After realizing this, I believe some renaming of functions is in order, so I'm going to do that work in a different branch and then resume when I can write the documentation to match functionality that *should* exist, but doesn't yet.