This commit represents my effort to consolidate all functions and variables on obvious names.  While doing that, I also created two primary loops that every model traversal uses.
Ingesting and validating data now uses a loop called ``import_loop``.  This name represents the attempt to take foreign data and import it into a Schematic.  The loop itself returns a dict with the convert data or throws an exception.

Exporting data, eg. writing out a serialization of the data, now works across ``export_loop`.  This loop used to be called ``apply_shape``, but I felt that was beginning to feel ambiguous in meaning.

I have also gone back to a convention of using ``to_*`` functions to imply the purpose of a serialization.  The function ``serialize()`` for example is now named ``to_primitive``.  There is also a ``to_native`` function which can be used to grab a dictionary of Python native types, which is useful for some database storage mechanisms.

I have not completed the renaming of all functions yet.  That will come in further commits.