Fix nested fields: set 'owner_model' and honor 'serialize_when_none'
This patch fixes an issue where `ListType` and `DictType` failed to
consult their inner fields on whether the nested items should be
serialized when None. The attribute on the container was applied
to the inner field as well.

Additionally, nested fields did not have their `owner_model` attribute
set and therefore `allow_none()` wouldn't have worked correctly
even if it were called. This is also fixed.