Change ListType validation messages to be a dict of index to message
Currently, ListType validation error messages do not tell you WHICH item
in the list was invalid; just that one of them was. This is problematic
for showing validation errors in a UI; you typically want to show the
message next to a specific field.

I believe that a dict of item index to one or more errors (in a list) is
the most usable format for UI developers. I'm also open to returning a
list of dicts, where the position in the list matches the position in
the input; though that would often necessitate empty placeholders.

Addresses: https://github.com/schematics/schematics/issues/307