Model differentiate valid from input data
Ideally _data will be an immutable dict and all writes happen in
_raw_data. Data access will return _raw_data first, then _data if set.
Serialization will always attempt to validate and output only if valid.

Another idea is to move the data state into a container, so that each
field can be individually queried, which may reduce the `context` needed
to be passed around during the transforms.