Model data state into custom ModelDict
This keeps the interface closer to the previous Model, while extracting
the data state handling into a single concern class.

This structure can be simplified in case the data state is stored with
the data itself, but will need more changes, specially with the types.
It is something we should investigate further.