Keep atoms `keys` param api and add tests
The keys params allows to select the resulting fields, which may be
useful for iterators that want to combine only specific data like field
name or type, or just get the values.

Since the use case is not consolidated, keep the test as reference of
this private API for now. But as a private API it is still open to changes.